checkIfWin() function first version
TODO: Consider refactoring this into the gameboard object instead. In that, I mean we store the winning array combinations on the gameboard object and then we use the row and column coordinates of a point to refer to the correct score map, and determine if that's a win or not.