Add massive bugfix for replaying / changing grid size and ending game when "best of" reached
..

was this even tested before we moved on to the next section? The game breaks when you change the grid size or if you try to replay.

For the grid size -- you didn't change the gameboard grid size variable when the input changed, so it was always stuck at 3.

For replaying -- child elements within #end-round-panel conveying a player victory persisted, which means a victory screen after the first round of the subsequent game.

For the best of reached, we just had to add a couple conditions that would check if any of the users' scores were the majority (half + 1); if so we call that a victory.